src/requirements.txt
moto
sqlalchemy
hdfs
kafka-python
minio
poyo
pytest
pytz
requests
apache-airflow
pandas<1.0.0
scikit-learn
joblib
pytest-spark
confluent-kafka
psycopg2-binary
typing_extensions
configparser<3.5.5
jinja2<2.11.0
werkzeug<1.0.0
bokeh